
\name{dnorm.truncated}
\alias{dnorm.truncated}
\title{Calculate density function for truncated normal random variables}
\description{Calculate density function for truncated normal random variables}
\usage{ dnorm.truncated(x, mu, sd, lower, upper, log)}
\arguments{
  \item{x}{value to calculate density at}
  \item{mu}{mean}
  \item{sd}{standard deviation}
  \item{lower}{lower truncation point}
  \item{upper}{upper truncation point}
  \item{log}{if TRUE, calculate the log density}
  }
\value{A vector of density values}
\author{Peter F. Craigmile}


