
\name{rnorm.truncated}
\alias{rnorm.truncated}
\title{Generate truncated normal random variables}
\description{Generate truncated normal random variables}
\usage{ rnorm.truncated(n, mu, sd, lower, upper)}
\arguments{
  \item{n}{number of variables to generate}
  \item{mu}{mean}
  \item{sd}{standard deviation}
  \item{lower}{lower truncation point}
  \item{upper}{upper truncation point}
  }
\value{A vector of truncated random variables }
\author{Christopher Hans and Peter F. Craigmile}
